A program that checks whether there are any unwanted parts inside the press machine.
For the press machine to operate properly, there must be no foreign objects inside. Otherwise, the machine could be damaged while in operation. To monitor this situation, I marked the press machine using a region of interest and trained an artificial intelligence model on the specified area. Due to privacy concerns, videos and photos cannot be shared.